Refactoring: use `strict_types`

+++ src/app/Http/Controllers/Api/V1/Search/VideoSearchController.php
@@ -31,7 +31,7 @@ class VideoSearchController extends Controller
             $this->videoRepository->search(
                 $query,
                 $request->get('page', 1),
-                $request->get('limit', config('settings.max_results_video')),
+                $request->get('limit', (int) config('settings.max_results_video')),
             ),
         );
